Carve MobileConnectionCoordinator out of MobileShellComposite (iOS refactor 5374)
Move the connection lifecycle into MobileConnectionCoordinator: pairing
entry (preview code, pairing URL, manual host), attach-ticket minting and
route selection, connect with the route/scoped-request fallback ladder,
reconnect-on-launch from the paired-Mac repository, disconnect/forget,
authorization-failure handling, connection-error localization, the
markMacConnection* health transitions, the workspace-list refresh task,
and ownership of the remote RPC client. The coordinator reaches the
facade's other pieces through the new MobileConnectionContext seam; the
composite keeps sign-in state, terminal input, remote create operations,
and exposes every public property as a forwarder so views and the feature
tests bind unchanged.

Behavior preserved: the pairingAttemptID and connectionGeneration UUID
guards move verbatim and stay hand-rolled. Pairing attempts are initiated
from callers' async contexts (SwiftUI actions, recovery, URL handlers),
so no single owned task's cancellation could replace the "is this still
the current attempt?" re-check after each suspension point; converting to
structured ownership is argued only for tasks a coordinator itself spawns.
Teardown ordering (replaceRemoteClient didSet -> stop listening, cancel
remote ops, reset output tracking) is identical via the context callback.

New package-level coverage: 9 MobileConnectionCoordinator tests for
manual-host validation, invalid pairing codes, ticket expiry before any
dial, paired-Mac persist + forget, reconnect-with-reminted-ticket,
unauthorized-connect reauth surfacing, cancelPairing, and sign-out reset.

Part of https://github.com/manaflow-ai/cmux/issues/5374
